+++ game/game.py
@@ -4,6 +4,11 @@ import pymunk.pygame_util
 import math
 import os
 
+from physics.world import criar_chao, criar_bloco
+from characters.boneco import criar_boneco
+from ui.menu import desenhar_menu
+from camera.controle import atualizar_camera
+from skins_system.skins import carregar_skin_pasta, carregar_thumbs
 
 def jogo(fila, config, gestos):
     
@@ -32,11 +37,12 @@ def jogo(fila, config, gestos):
     conexao_garra = None
     ponto_arrastando = None
     mouse_joint = None
+    modo_edicao = False
 
     rodando = True
 
     # SPRITES
-    sprites_boneco = {}
+    sprites_boneco = carregar_skin_pasta("teste")
 
     skins_disponiveis = (os.listdir("skins")
                          
@@ -44,40 +50,6 @@ def jogo(fila, config, gestos):
         else [])
 
     thumbs_skins = {}
-
-    # FUNÇÕES AUXILIARES
-    def criar_chao():
-
-        chao = pymunk.Segment(espaco.static_body, (0, ALTURA - 50), (LARGURA, ALTURA - 50), 5)
-
-        chao.friction = 1.0
-
-        espaco.add(chao)
-
-    def criar_bloco(x, y, largura, altura, massa=10):
-
-        momento = pymunk.moment_for_box(massa, (largura, altura))
-
-        corpo = pymunk.Body(massa, momento)
-
-        corpo.position = x, y
-
-        forma = pymunk.Poly.create_box(corpo, (largura, altura))
-
-        forma.friction = 0.5
-        forma.elasticity = 0.3
-
-        espaco.add(corpo, forma)
-
-        return forma
-
-    def criar_corda(corpo_a, corpo_b, ancora_a, ancora_b, comprimento):
-
-        corda = pymunk.SlideJoint(corpo_a, corpo_b, ancora_a, ancora_b, 0, comprimento)
-
-        espaco.add(corda)
-
-        return corda
 
     # PONTOS CONTROLE
     pontos_controle = []
@@ -98,105 +70,14 @@ def jogo(fila, config, gestos):
 
         pontos_controle.append(corpo)
 
-    # BONECO
-    def criar_boneco():
-        torco = criar_bloco(LARGURA // 2, 200, 75, 90)
-        cabeca = criar_bloco(LARGURA // 2, 90, 80, 80)
-        bresq = criar_bloco(LARGURA // 2 - 50, 200, 25, 50)
-        antesq = criar_bloco(LARGURA // 2 - 50, 275, 25, 50)
-        bradir = criar_bloco(LARGURA // 2 + 50, 200, 25, 50)
-        antdir = criar_bloco(LARGURA // 2 + 50, 275, 25, 50)
-        peresq = criar_bloco(LARGURA // 2 - 25, 325, 25, 50)
-        panesq = criar_bloco(LARGURA // 2 - 25, 400, 25, 50)
-        perdir = criar_bloco(LARGURA // 2 + 25, 325, 25, 50)
-        pandir = criar_bloco(LARGURA // 2 + 25, 400, 25, 50)
-        cintura = criar_bloco(LARGURA // 2, 250, 75, 35)
-
-        # CORDAS
-        criar_corda(torco.body, cabeca.body, (0, -45), (0, 40), 25)
-        criar_corda(torco.body, bresq.body, (-37.5, -30), (0, -25), 20)
-        criar_corda(bresq.body, antesq.body, (0, 25), (0, -25), 25)
-        criar_corda(torco.body, bradir.body, (37.5, -30), (0, -25), 20)
-        criar_corda(bradir.body, antdir.body, (0, 25), (0, -25), 25)
-        criar_corda(torco.body, cintura.body, (0, 45), (0, -17.5), 25)
-        criar_corda(cintura.body, peresq.body, (-25, 17.5), (0, -25), 25)
-        criar_corda(peresq.body, panesq.body, (0, 25), (0, -25), 25)
-        criar_corda(cintura.body, perdir.body, (25, 17.5), (0, -25), 25)
-        criar_corda(perdir.body, pandir.body, (0, 25), (0, -25), 25)
-
-        # Corda
-        criar_corda(pontos_controle[2], cabeca.body, (0, 0), (0, -40), 100)
-        criar_corda(pontos_controle[0], antesq.body, (0, 0), (0, 25), 150)
-        criar_corda(pontos_controle[4], antdir.body, (0, 0), (0, 25), 150)
-        criar_corda(pontos_controle[1], peresq.body, (0, 0), (0, 25), 400)
-        criar_corda(pontos_controle[3], perdir.body, (0, 0), (0, 25), 400)
-
-        return{
-            "cabeca": cabeca,
-            "torco": torco,
-            "bresq": bresq,
-            "antesq": antesq,
-            "bradir": bradir,
-            "antdir": antdir,
-            "peresq": peresq,
-            "panesq": panesq,
-            "perdir": perdir,
-            "pandir": pandir,
-            "cintura": cintura
-        }
-
-    # SKINS
-    def carregar_skin_pasta(nome_pasta):
-
-        nonlocal sprites_boneco
-
-        caminho = f"skins/{nome_pasta}"
-
-        tamanhos = {
-            "cabeca": (80, 80),
-            "torco": (75, 90),
-            "bresq": (25, 50),
-            "antesq": (25, 50),
-            "bradir": (25, 50),
-            "antdir": (25, 50),
-            "peresq": (25, 50),
-            "panesq": (25, 50),
-            "perdir": (25, 50),
-            "pandir": (25, 50),
-            "cintura": (75, 35)
-        }
-
-        sprites_boneco = {}
-
-        for parte, tamanho in tamanhos.items():
-
-            arquivo = f"{caminho}/{parte}.png"
-
-            if os.path.exists(arquivo):
-
-                img = pygame.image.load(arquivo).convert_alpha()
-
-                sprites_boneco[parte] = pygame.transform.scale(img, tamanho)
-
-    def carregar_thumbs():
-
-        for pasta in skins_disponiveis:
-
-            caminho = f"skins/{pasta}/cabeca.png"
-
-            if os.path.exists(caminho):
-
-                img = pygame.image.load(caminho).convert_alpha()
-
-                thumbs_skins[pasta] = pygame.transform.scale(img, (60, 60))
-
     # CENÁRIO
-    criar_chao()
-    meu_boneco = criar_boneco()
+    criar_chao(espaco, LARGURA, ALTURA)
+
+    meu_boneco = criar_boneco(espaco, pontos_controle, LARGURA)
     
-    carregar_thumbs()
-    criar_bloco(
-        LARGURA // 2 + 200, 100, 20, 100, 50)
+    thumbs_skins = carregar_thumbs(skins_disponiveis)
+    
+    criar_bloco(espaco, LARGURA  // 2 + 200, 100, 20, 100, 50)
 
     # MOUSE
     mouse_body = pymunk.Body(body_type=pymunk.Body.KINEMATIC)
@@ -207,29 +88,8 @@ def jogo(fila, config, gestos):
         tela.fill((240, 240, 240))
 
         mouse_pos = pygame.mouse.get_pos()
-
-        # CAMERA
-        if fila is not None and not fila.empty():
-
-            try:
-
-                while not fila.empty():
-                    pontos = fila.get_nowait()
-
-                if not modo_edicao and pontos and len(pontos) >= 21:
-
-                    escala_x = 2
-                    escala_y = 1.5
-
-                    pontos_controle[0].position = (pontos[4][0] * escala_x, pontos[4][1] * escala_y)
-                    pontos_controle[1].position = (pontos[8][0] * escala_x, pontos[8][1] * escala_y)
-                    pontos_controle[2].position = (pontos[12][0] * escala_x, pontos[12][1] * escala_y)
-                    pontos_controle[3].position = (pontos[16][0] * escala_x, pontos[16][1] * escala_y)
-                    pontos_controle[4].position = (pontos[20][0] * escala_x, pontos[20][1] * escala_y)
-
-
-            except Exception as erro:
-                print("Erro camera:", erro)
+        
+        atualizar_camera(fila, pontos_controle, modo_edicao)
 
         # GESTOS
         acao = "Parado"
@@ -399,23 +259,7 @@ def jogo(fila, config, gestos):
             pygame.draw.circle(tela, (0, 100, 255), (int(p.position.x), int(p.position.y)), 15, 2)
 
         # MENU
-        pygame.draw.rect(tela, (150, 150, 150), rect_abrir_menu)
-
-        tela.blit(
-            fonte.render("ABRIR MENU", True, (255, 255, 255)), (LARGURA - 140, 20))
-
-        # MENU ABERTO
-        if menu_aberto:
-
-            fundo_menu = pygame.Rect(LARGURA - 260, 60, 250, 300)
-
-            pygame.draw.rect(tela, (40, 40, 40), fundo_menu)
-
-            pygame.draw.rect(tela, (255, 255, 255), fundo_menu, 2)
-
-            texto = fonte.render("MENU ABERTO", True, (255, 255, 255))
-
-            tela.blit(texto, (LARGURA - 220, 80))
+        desenhar_menu(tela, fonte, rect_abrir_menu, menu_aberto, LARGURA)
         
         # CURSOR CAMERA
         pygame.display.flip()
